% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/readFolder_function.R
\name{readFolder}
\alias{readFolder}
\title{Read folder}
\usage{
readFolder(dir, pattern = "*", sep = "\\t", format = "data.frame",
  header = T, fun)
}
\arguments{
\item{dir}{directory to folder.}

\item{pattern}{wildcard pattern for filenames. Defaults to "*".}

\item{sep}{separator for columns in data files. Defaults to tabulator.}

\item{format}{the format to be returned. Either a list or data.frame. Defaults to data.frame.}

\item{header}{does the files contain a header. Defaults to TRUE.}

\item{fun}{function to be applied to each file in the folder.}
}
\description{
This function reads all files in a folder.
}
\examples{
Calculate the cumulative value for a variable("score") in a new variable("cumscore")
cumulativeSum = function(data) {
   data[, "cumscore"] = cumsum(data[, "score"])
   return(data)
}
dir = "c:/experiments/experiment1/"
data = readFolder(dir, fun = cumulativeSum)
}

